django_test/requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.0.14
django-admin-interface==0.30.0
django-colorfield==0.14.0
mod_wsgi==5.0.2
mssql-django==1.5
mysqlclient==2.2.7
pillow==11.1.0
pyodbc==5.2.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.2
reportlab==4.3.1
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.1
